imagem_para_pdf_ferramenta/requirements.txt
Flask==2.3.2
Pillow==9.3.0
PyPDF2==3.0.0
